requirements.txt
Django==4.2.4
djangorestframework==3.14.0
python-dotenv==1.0.0
drf-yasg==1.21.7
pytest==7.4.1